Island Escape 1.1
Fixed a 'bug' with the lunge and removed the drag from the rigid body. Created a based prefab for the guard.  Two guard types, one that shoots at the player, and one that chases the player. Both guards have a range the player needs to be in for this functionality.